nvcomp_minimal/pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "Cython"]
build-backend = "setuptools.build_meta"

[project]
name = "nvcomp-minimal"
version = "0.1.0"
description = "Minimal Cython wrapper for nvcomp zstd decompression"
requires-python = ">=3.12"
dependencies = [
    "cupy-cuda12x",
    "numpy",
]

[project.optional-dependencies]
dev = [
    "numcodecs",
    "nvtx",
]

# [tool.setuptools]
# ext-modules = [
#   {name = "nvcomp_minimal", sources = ["nvcomp_minimal/zstdt .pyx"]}
# ]
